schema-miers/requirements.txt
# include all packages and versions needed for pipeline for staging:
# python-dotenv>=0.21.0
# requests>=2.28.1
# dbt-core>=1.5.0
# dbt-snowflake>=1.5.0
# can also contain packages from Nexus, for staging:
bi-snowflake-connector==1.2.2
awswrangler==3.5.2
boto3==1.34.48
pyarrow==10.0.1
xlrd==2.0.1
paramiko==3.4.0
python-dotenv==1.0.1
fabric==3.2.2
pandas==2.2.0
pytz==2024.1
openpyxl==3.1.1
dbt-snowflake>=1.7.0,<1.8.0
dbt-core>=1.7.0,<1.8.0
